project/jwt_tokens.py
from django.shortcuts import redirect
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework_simplejwt.views import TokenObtainPairView
from project.data_scraper import scrape_data

class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
    
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        # Add custom claims
        token['username'] = user.username
        token['password'] = user.password

        # ...

        return token

class MyTokenObtainPairView(TokenObtainPairView):
    serializer_class = MyTokenObtainPairSerializer


def scrape_data_view(request):
    country = request.GET.get('country')
    contents = request.GET.get('contents')
    scrape_data(country, contents)
    return redirect('http://127.0.0.1:8001/')